+++ examples/all.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 
+../bin/a11y-contrast.js black-white.json
 ../bin/a11y-contrast.js ibm-v1.json
 ../bin/a11y-contrast.js ibm-v2.1.json --min-ratio-3=40 --min-ratio-4.5=50 --min-ratio-7=70
 ../bin/a11y-contrast.js open-color.json